Fixup $ref resolution and use correct $id property
jsonschema Draft-06 changed "id" to "$id" to avoid collisions with "id"
being used in a schema. Update everything to use $id so that the schema
files are compliant with draft-06.

At the same time, normalize the namespace for devicetree.org schema and
metaschema files. With this change, all devicetree schemas are under the
http://devicetree.org/ URL and the code is adjusted to load
devicetree.org schemas from the local filesystem instead of insisting on
HTTP. (The code for doing this is probably suboptimal, but it can be
fixed up to be cleaner later)

Signed-off-by: Grant Likely <grant.likely@arm.com>