meta-schema: add meta-schema for compatible descriptions
'compatible' is ultimately just an array of strings in a defined order.
Descriptions can be complicated with logical expressions of valid
combinations.

Expressing these different conditions means schema structure may be at
differing levels. We use the trick of "$ref: '#'" which basically makes
the parsing of schema recursive.

Some of this can probably be generalized for other string arrays, but that
can be done when we have multiple cases.

Signed-off-by: Rob Herring <robh@kernel.org>