meta-schema: rework properties handling to group properties in different files
Putting all properties into core.yaml won't scale. Use 'allOf' list so
we can at least group properties together in different files. We lose
the ability to warn on unknown properties, but the consensus in the json
schema community seems to be doing that should be a validator option.

If this becomes unmanageable, then perhaps we can generate the schema
with the list of files with properties.

Signed-off-by: Rob Herring <robh@kernel.org>