Add source line & column numbers to error messages
Errors messages are more useful when they can be matched to a line of
source code. Add some basic resolving of the file, line and column
information to error messages and output in a format that Vim
understands (filename:<line>:<column>: error message).

This feature depends on the switch to ruamel.yaml from the previous
commit. It also factors out some of the error formating code into the
common dtschema library so it can be used by other tools.

Signed-off-by: Grant Likely <grant.likely@arm.com>