schemas/dt-core: avoid recursing into nodes
The dt-validate script already recurses into nodes to select schema and the
core schema is always selected. So there is no need for the schema itself
to descend into subnodes too. Doing it in the schema also makes for vague
error messages due to the use of 'oneOf'. Using best_match() method can
help on that.

Signed-off-by: Rob Herring <robh@kernel.org>