dt-validate: fixup schema for array sizes and scalar values
For DT schema, we want to default to array sizes and values being fixed and
known. Doing this with json-schema requires having 'minItems', 'maxItems',
and 'additionalItems' properties. That gets to be a lot of redundant
boilerplate in binding schema, so let's generate those properties instead.

The DT yaml encoding always encodes integer values as a matrix regardless
of how many values a property can or does have. For example, '#size-cells'
may have a value such as '[[ 1 ]]'. This too would end up being a lot of
boilerplate defining an array of arrays. So fixup single value schemas
so we can just specify the single or set of values allowed.

The fixup is done when check_schema is called. Therefore, calling
check_schema before validation is required and the meta-schema remains
written for the un-modified schema.

Signed-off-by: Rob Herring <robh@kernel.org>

move schema fixup