meta-schemas: rework required schema properties for cell based properties
For properties which can be an array, but are defined to have a single
entry only, binding schemas only need to specify 'maxItems' is 1.

For arrays, require 'items' and drop requiring 'minItems' and 'maxItems'
as we handle these implicitly by default.

Also, use 'propertyNames' as this is the preferred way to restrict allowed
properties instead of 'additionalProperties'.

Signed-off-by: Rob Herring <robh@kernel.org>