dt-validate.py: rework validator caching
When there are a large number of schema that never match any nodes, then
creating a dtschema.DTValidator for every schema is a waste of time, so
only create the object when the schema is selected.

Also, move $select_validator to check_trees instead of load_binding_schema
in preparation to dump the processed schema out to yaml.

Signed-off-by: Rob Herring <robh@kernel.org>