schemas: split interrupt related schema from core into its own file
The core schema will grow unwieldy if we keep adding common bindings to it,
so start with moving the interrupts binding out.

Note the 'allOf can probably be removed and properties put at the top
level, but that currently doesn't work with the fixups. This is a way to
prevent any fixups.

Signed-off-by: Rob Herring <robh@kernel.org>