dtschema: Rework $nodename property handling
Adding the $nodename property was not working if only 'patternProperties'
was present. The property was also not getting represented the same way
as other string properties. Instead of adding $nodename to properties,
let's generate the 'select' schema if $nodename is present in properties.
That eliminates one of the main 'select' patterns.

Signed-off-by: Rob Herring <robh@kernel.org>