Add support for custom keyword 'typeSize'
DT properties can have 8, 16, 32, or 64 bit sized values and arrays of
values. The size information is represented as a YAML tag in the DT data.
In order to validate the sizes of properties, a custom validation keyword
is needed as json-schema doesn't provide a way to do this. The closest
way would be using 'maximum', but that applies to integers and we need a
schema for arrays. Add 'typeSize' as a schema keyword for arrays which
has the size in bits for the array elements.

Signed-off-by: Rob Herring <robh@kernel.org>