schemas: fix types checks for elements beyond the 1st element
In json-schema, 'items' can be a list of schemas corresponding to each
element in a list or can be a schema for all elements in a list. The
former case was being used, but that results in only the 1st list element
being validated. For example, this data would pass:

[ !u8 [ 0, 1, 2, string ] ]

Convert all the defined types to have a schema for 'items' rather than a
list. The one exception is phandle+args where the elements are different
schema. Also, ensure the type is array and minItems and maxItems are
explicitly set.

Also extend uint32 with matrix type and support different encodings as
the source format can vary.

Signed-off-by: Rob Herring <robh@kernel.org>