schemas: extend the allowed property and node names to 64 characters
The DT spec says the max node and property name length is 31 characters,
but this has never been enforced nor followed. The consensus seems to be
to change the spec, but that has not happened yet. For now, just bump the
limit to 64 characters which should be enough for known cases.

Signed-off-by: Rob Herring <robh@kernel.org>