Improve validation speed making line numbers optional
Loading data for validation is slow as it uses the roundtrip loader to
preserve line numbers. The speed is a factor if validating lots of data
files with only a few schema. Using the safe loader is 2-3x faster. The
line numbers aren't too useful often because the input data file is an
intermediate file.

Add a command-line option '-n' (like grep) to enable line numbers.

Signed-off-by: Rob Herring <robh@kernel.org>