dtschema: Add a $ref resolution check
Currently, $ref values don't have to be valid to pass meta-schema checks
and they don't fail until the schema is applied. Add a check method to
walk a schema and resolve any $ref.

This could be improved by printing line numbers and integrating with the
meta-schema validation, but it's good enough for now.

Signed-off-by: Rob Herring <robh@kernel.org>