dtschema: add a schema fixup for int arrays
Arrays are formatted as '[ !u8 [ 0, 1, 2 ] ]', but we want to hide that
complexity from the schema to just define the array size and constraints
of the elements.

So a schema like this:

array-prop:
  minItems: 1
  maxItems: 2
  items:
    enum: [ 0, 1, 2 ]

needs to be expanded to:

array-prop:
  items:
    minItems: 1
    maxItems: 2
    items:
      enum: [ 0, 1, 2 ]

This is only done when there's an 'items' schema with integer constraints
as some properties (reg, interrupts, etc.) need to define the outer array
size.

Signed-off-by: Rob Herring <robh@kernel.org>