dtschema: add implicit properties to schemas
Any node can have a 'phandle' property automatically added if it is
referenced. 'interrupt-parent' is also valid in any node with
'interrupts' property. 'status' is valid for any node.

Add these in the schema fixup processing so that schemas with
'additionalProperties: false' will not raise an error for these
properties.

Signed-off-by: Rob Herring <robh@kernel.org>