dtschema: Deal with the disabled nodes
When we have a disabled nodes, we don't want to validate that all the
required properties are there, since some might be external resources
usually provided by a DTS.

However, boards DTS that don't enable that component will not be able to
provide that resource and would fail to pass the required properties check.

We'll still want to validate the existing properties though.

In order to work around it, let's just ignore any error message mentionning
a missing required property on a disabled node.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>