dtschema: Deal with the disabled nodes
When we have a disabled nodes, we don't want to validate that all the
required properties are there, since some might be external resources
usually provided by a DTS.

However, boards DTS that don't enable that component will not be able to
provide that resource and would fail to pass the required properties check.

We'll still want to validate the existing properties though.

In order to work around it, let's generate two schemas from a file input:
 * the first one will match on nodes that don't have a status property set
   to disabled, using a refined select statement.
 * the second one will match on nodes that have a status property that
   is set to disabled, and we'll drop the required section entirely. That
   will allow to validate the provided properties for any given board on a
   disabled node, while not generating any spurious missing property
   message.

Fixes issue #8.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>