dtschema: Add pinctrl properties by default
Since pinctrl properties are handled by the core itself most of the time,
and are not always essential to the operations of a driver (for example if
the bootloader is in charge of the muxing), we would have to add the
pinctrl-names and pinctrl-[0-255] properties to pretty much every YAML
schema.

Add them in the tools instead just like status, but only do so if there's
no pinctrl-related properties defined.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>