test: split test case data into pass and fail files
The use of good/bad node names doesn't work if schema require a certain
node name. So split the tests into separate files for expected pass and
expected fail.

Descending the tree doesn't work well either because we may expect a
schema with child nodes to fail, but if we iterate to the child node there
may not be any failure. Instead, only iterate over the top-level nodes.

Signed-off-by: Rob Herring <robh@kernel.org>