dtschema: Fixup schemas under allOf conditionals
The Draft-07 specification introduced the if then and else conditionals in
schemas.

However, if multiple conditionals are to be used, allOf needs to be used to
express them.

In such a case, the schemas fixup wouldn't run on the schemas found under
the conditionals, which led to unnecessary boilerplate. Fix this.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>