meta-schemas: interrupts: allow interrupt-controller nodes to be objects
RISC-V cores contain basic interrupt controllers, described in
DT 'cpus' data in the form:

        cpu@0 {
		[... other CPU properties ...]
	        interrupt-controller {
                        #interrupt-cells = <1>;
                        compatible = "riscv,cpu-intc";
                        interrupt-controller;
                };
        };

Since the existing interrupts meta-schema requires the type of the
"interrupt-controller" node to be a boolean, it doesn't support this
usage.  Support this by relaxing the type constraint on
"interrupt-controller" nodes to also accept objects.  Since this usage
does not require an "#interrupt-cells" node at the same tree depth as
the "interrupt-controller" object, remove the dependency accordingly.

Signed-off-by: Paul Walmsley <paul.walmsley@sifive.com>
Cc: Rob Herring <robh@kernel.org>
Signed-off-by: Rob Herring <robh@kernel.org>