schemas: types: Introduce a non-unique string array type
So far, the string-array type mandates that all the items are not unique.

However, a good number of string properties allow for non-unique strings
(gpio-line-names, pinctrl-names, dma-names, etc), and this is pretty
difficult to disable the unique check, which basically requires to redefine
the type all over again.

In order to ease the definitions of these properties, let's add a new type
that allows for string arrays with non-unique values.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>