schemas: core: Change *-names default to allow unique strings
So far, the core schema mandates that all the *-names properties contain
strings that are unique.

However, a good number of the properties ending with names allow for
non-unique values (gpio-line-names, pinctrl-names, dma-names, etc).

In order to have a proper behaviour in those case, let's make the core
schema allow for non-unique strings, and have the properties that don't
allow it explicitly define it by using the old default string-array type
directly.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>