meta-schema: Fix 'interrupt-controller' meta-schema
Commit 379741ead1e6 ("meta-schemas: interrupts: allow
interrupt-controller nodes to be objects") breaks 'interrupt-controller'
schema because both clauses are true since a boolean schema is also an
object (schemas are objects). What is desired is the schema should
define the data to be an object (i.e. a child node). Add a meta-schema
for defining child nodes and reference that instead.

Fixes: 379741ead1e6 ("meta-schemas: interrupts: allow interrupt-controller nodes to be objects")
Signed-off-by: Rob Herring <robh@kernel.org>