meta-schemas: Allow cell arrays to be described as a single entry
Currently we're mixing array and scalar properties in schemas. The tools
fix-up the schema, but it doesn't look like valid json-schema. Allow for
single element arrays to have an implied length of 1 and just describe
them with scalar properties.

For now, both forms are allowed. Once we fix cases using the old way, the
meta-schema can be updated to disallow the old way.

Signed-off-by: Rob Herring <robh@kernel.org>