dtschema: Only fixup the array size when 'items' is a list
When 'items' is schema rather than a list, then it applies to all items.
If the array sizes are not fixed, then it also allows either source
encoding '<1 2 3 4>' or '<1>, <2>, <3>, <4>' to be valid.

However, once an array size is defined, then the former encoding must be
used (or we'll need to extend the fixups some more).

Signed-off-by: Rob Herring <robh@kernel.org>