Add a check for unnecessary quoting of values
Generally, quoting strings in YAML is not necessary. There are a few
exceptions where YAML requires quoting and generally it is only for
special characters at the beginning of values. The only case we care about
is '#' at the beginning. There are a few other allowed, but not recommended
characters we don't currently handle.

Currently, this only checks 'const' values and string lists (enum).

Signed-off-by: Rob Herring <robh@kernel.org>