dtschema: Don't use default $nodename for select
The current code adds a default $nodename property if it's not found in the
processed schema, and then calls the fixup functions.

Amongst those functions, add_select_schema will test whether $nodename is
found. If it is, then it will generate a select statement based on the
$nodename value, and if not, will generate a select statement set to false.

However, since we added the default before calling add_select_schema, that
latter condition never occurs, which can lead to schemas without a
compatible, $nodename or an explicit select statement being always selected
instead of never.

Such a case can happen with the recently merged in Linux 5.3 panel-common
schema.

Instead, only use $nodename for select if $nodename is something other
than 'true'.

Reported-by: Maxime Ripard <maxime.ripard@bootlin.com>
Signed-off-by: Rob Herring <robh@kernel.org>