meta-schemas: Allow to use unevaluatedProperties
The draft-08 specification will introduce a keyword to express that we
don't want any properties that have not been validated, as opposed to
rejecting any property that haven't been described in the same schemas.

This will hopefully allow us to deal with generic vs specific bindings a
bit nicer, so we want to introduce it ahead of time, even though it's not
used yet.

Signed-off-by: Maxime Ripard <mripard@kernel.org>