dt-validate: Fix unmatched schema reporting
The unmatched node printing had some errors. It's not sufficient to just
check the node has 'compatible' for whether there was a match on
compatible. We must ensure both 'select' uses 'compatible' to match and
the node has 'compatible' to determine if there was a match or not. In
particular, as there is a schema for all root nodes, root nodes without
a schema for their compatible were never flagged as missing.

Fixing the root node matching revealed that the root path was not
printing correctly, so fix that.

Extend the '-m' option to be repeatable so printing unmatched
compatibles and no match at all can be enabled separately. If '-m' is
given twice, then also print nodes without any matching schema.

Signed-off-by: Rob Herring <robh@kernel.org>