meta-schemas: Exclude some patterns from vendor specific properties
Some common property names match on the vendor specific property
meta-schema, so we need to exclude them. This is a little tricky to do.
For any matching patterns, we need the schema to be false so that the
'then' condition is not evaluated. This is because a property name which
doesn't match a pattern also validates as true.

Signed-off-by: Rob Herring <robh@kernel.org>