meta-schemas: string-array rework
The string-array meta-schema was flawed in that only the 1st level
property schema names were getting checked. This meant any property name
under 'oneOf' was allowed for example.

To fix this, we need to restructure the schema so we can distinguish 1st
level properties from when we recurse the schema. With the property names
correctly checked, add 'contains', 'allOf', 'deprecated', and '$ref' to
the property list.

Signed-off-by: Rob Herring <robh@kernel.org>