dtschema: Print the sub-schema failures under conditionals
The errors 'is not valid under any of the given schemas' is not too useful,
but is common with 'allOf', 'anyOf', and 'oneOf' schemas. In these cases,
ValidationError.context will be populated with sub-schema errors. Printing
them all is not too helpful either. Let's print the one selected by
jsonschema lib best_match() method and the ones with a sub-schema path as
they tend to point in the right direction.

Signed-off-by: Rob Herring <robh@kernel.org>