dtschema: Rework int array to matrix fixups
Using _is_int_schema to match int arrays doesn't work if we have a list
of 'items' with just descriptions as that looks the same as a string or
phandle list. To fix this, let's also look for a $ref to a 'uint*-array'
type in addition to integer keywords. As part of this, _fixup_int_array
becomes redundant and can be removed.

The items size fixup needs to be done after all the other fixups and after
all the recursion. Otherwise, we miss some instances.

Signed-off-by: Rob Herring <robh@kernel.org>