meta-schemas: Add a recursive keywords meta-schema
Currently, schemas are not getting checked at all levels for valid
json-schema keywords. The default behavior is to ignore unknown keywords,
but that is not desired in our case. Add a new meta-schema which recurses
into the sub-schemas.

Restructure things a bit to have a recursive schema that just checks for
valid json-schema and any global restrictions we have. Then core.yaml is
mostly schemas for specific properties.

Signed-off-by: Rob Herring <robh@kernel.org>