meta-schemas: Require 'additionalProperties' or 'unevaluatedProperties'
While the default behavior is additional/unevaluated properties are
allowed, there's not any hard pattern we can match to fixup/enforce the
non-default behavior (e.g. additionalProperties: false). This is a
constant source of review issues, so an automated check is needed.

If we have a $ref, then either 'additionalProperties' or
'unevaluatedProperties' is required. If there's no $ref, then
'additionalProperties' is required.

Signed-off-by: Rob Herring <robh@kernel.org>