clocks: Allow nodes being called "clocks"
Even though this is less ideal, "clocks" is both used as a generic
property name in many nodes, but also as a node name (mostly to group
multiple clocks as child nodes).
In the generic clock binding we acknowledge this and allow "clocks" to
be of type "object" as well, but fail to do something similar in the
meta-schemas.

Allow "clocks" being a node, as an alternative to the phandles array, to
let DTs using "clocks" as a node name pass the tests.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>