dtschema: Fixup '$ref' which need to be under an 'allOf'
In json-schema draft7 and earlier, any other schema keywords at the same
level as a $ref are ignored. This changes in the 2019-09 release. The
old way around this was to put '$ref' under an 'allOf'. This is not
obvious and a common issue reviewing schema. Let's fix this up
automatically so schema can be written in 2019-09 compliant way and we
can solve the review issue.

Signed-off-by: Rob Herring <robh@kernel.org>