tools/dt-extract-example: support examples with root node
Examples are prefixed with boiler plate in the normal case.
In the rare case a root node is required this boiler plate shall not be
used.

Example with root node:

    / {
        chosen {
            ...;
        }
    }

Teach the script to recognize that the example starts with a root node
and avoid the boiler plate in such cases.
The root node is recognized using a simple regular expression search.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>