dt-validate: Check status of correct instance
When deciding whether or not to emit an error, look at the 'status'
property of the instance that failed to validate rather than of the
node that is being checked.

Cases where this can happen are when a schema describes required
properties for patternProperties or other child nodes, in which case
the invalid instance differs from the node being checked.

Signed-off-by: Thierry Reding <treding@nvidia.com>