dt-validate: Print the schema file causing the error
Since multiple schemas can be applied to a node, it's not always clear
where the source of an error was. The jsonschema errors also don't have
any way to get back to the schema file, but we have at least the starting
schema file, so let's print it.

If the schema file references other files, then those files could be the
actual source. There's not currently a way to track $refs, so we only get
the starting point.

Signed-off-by: Rob Herring <robh@kernel.org>