schemas: iio: Drop io-channel-ranges
'io-channel-ranges' is defined to be a consumer property, but it is
frequently defined on the provider side which has no effect (at least
for Linux). It's unlikely that we'd have a case where we really have a
bunch of child nodes sharing an io-channel and it's too burdensome to
just use 'io-channels' in each child node. So let's just drop
'io-channel-ranges'.

Signed-off-by: Rob Herring <robh@kernel.org>