Remove 'allOf' when used with '$ref'
In json-schema 2019.09 syntax, keywords in addition to an $ref are no
longer ignored so an 'allOf' is no longer needed. We fix this up when
preprocessing the schemas until jsonschema module has 2019.09 support.

Remove all the cases in the schemas and test schemas so others don't
copy the old syntax.

Signed-off-by: Rob Herring <robh@kernel.org>