dtschema: Add assigned-clock* properties by default
Since the assigned-clocks and related properties are handled by the core
itself most of the time, can be applied to pretty much any device node
and can depend on a specific board setup, we would eventually have to
add these properties to pretty much every YAML schema.

Add them in the tools just like status and pinctrl properties, but only
if the schema doesn't define them already and if the node has a clocks
property.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Rob Herring <robh@kernel.org>