dtschema: Rework int array detection and fixups
Currently, the int array detection relies on finding a type $ref. That was
because we couldn't distinguish an int array from a string array or matrix
if the items had no constraints. This is not ideal especially for
properties where we know the type. These are primarily ones with a unit
suffix. For now, let's detect these by looking at the property name. This
isn't really ideal, but should handle most cases.

Besides enabling fixups on properties with unit suffixes, this change has
the side effect of allowing both array forms in some additional cases.

Signed-off-by: Rob Herring <robh@kernel.org>