dtschema: Always disable duplicate keys
Duplicate keys were allowed for YAML encoded DT files due to some
collisions with node names and property names. This is valid, but
something discouraged and has undefined behavior. The spec should
probably be updated to disallow it.

If we end up needing to add this back, we should instead warn and
then fallback to allowing duplicates.

Signed-off-by: Rob Herring <robh@kernel.org>