dtschema: Add 'description' from schemas to error messages
Some error messages don't give any clues where they come from or how to
fix them. To give some hints, let's extract the lowest 'description'
and '$id' in the schema path. This is complicated since the description
could be in 'error.schema' or above that in the parent schema where we
have to walk the schema from the root.

Signed-off-by: Rob Herring <robh@kernel.org>