dtschema: Drop forcing YAML 1.1 parsing
This reverts commit 1c7f33244d91 ("dtschema: Force the loader to use
YAML 1.1"). The ruamel C parser has supported YAML 1.2 since 2018 which
was well before that commit. It seems like it was likely a configuration
issue or need to adjust dependencies. In any case, using YAML 1.1 is
actually problematic for a few cases where 'on' and 'off' are used in
binding and get treated as boolean instead of strings.

Signed-off-by: Rob Herring <robh@kernel.org>