dtschema: Make the list of schemas a dict instead
As we need to retrieve schemas by $id, it makes more sense to use a
dict with $id as the key instead of a list to store the schemas.

Since the dt-mk-schema output is also the same format, we need to be
able to read old format which is easy to detect and translate.

Signed-off-by: Rob Herring <robh@kernel.org>