Add support to validate dtbs
There's a desire to be able to run DT validation on running systems
where there might not be an original source dts file. As the YAML DT
format is tied to the DT source level annotations (e.g. /bits/,
brackets, etc.), a dtb->dts->yaml or dtb->yaml conversion won't work.

Therefore, a different approach is used. pylibfdt is used to unflatten
a dtb directly into python data structures. Then type information
from the schemas are used to decode the property values into the right
types.

This approach removes the YAML format from the picture making the format
for validation purely an internal format (though the dtb2py program
will dump it) and avoids some short comings of the YAML format. This
change also improves the validation some. For example, some validation
could pass before with incorrect bracketing because the validation is
not aware of cell sizes.

Signed-off-by: Rob Herring <robh@kernel.org>