dtschema: Strip 'additionalProperties: true' from schemas
An 'additionalProperties: true' makes all properties evaluated, so an
'unevaluatedProperties: false' schema doesn't work when referencing a
schema with 'additionalProperties: true'.

Really, 'additionalProperties: true' should be dropped, but it is used as
a flag for common schemas which are usually incomplete and included by
device specific schemas. Removing it would mean dropping the meta-schema
requiring either 'additionalProperties' or 'unevaluatedProperties'.

Signed-off-by: Rob Herring <robh@kernel.org>