meta-schemas: Fix check for json-schema vocabulary used as a DT property name
The existing check for json-schema vocabulary used as a DT property name
fails to recurse into sub schema such as:

allOf:
  - if:
      properties:
        ...

Move this check to keywords.yaml which is a more logical place for it,
always gets applied (some schemas only use base.yaml), and recurses the
schemas fully. We can also remove the duplicated list of allowed json-schema
vocabulary.

Signed-off-by: Rob Herring <robh@kernel.org>