test: child-node-fail: Fix 'reg' entry sizes
The bad 'reg' size depends on the brackets to be the wrong number of
entries, but since the entry size was wrong with only 1 cell the overall
size was correct (1 address cell and 1 size cell). Update the entry to
be 2 entries which is too many and will fail tests regardless of source
brackets.

Signed-off-by: Rob Herring <robh@kernel.org>