dt-validate: Add '-l' option to limit schemas for validation
In order to do validation on dtb files, a complete set of schemas is
always needed to extract type information. That doesn't work when only
validating with a subset of schemas. In order to continue to support
that use case, add an explicit option, -l or --limit, to dt-validate to
provide a specific schema to use for validation. The option can be a
substring which is useful for selecting a number of schemas.

Signed-off-by: Rob Herring <robh@kernel.org>