schemas: i2c-controller: Add 'no-detect' property
In some cases, operating systems may attempt to detect the presence of
devices on an I2C bus by, for example, attempting to read specific
registers from a list of addresses which some common peripherals is
known to use.

Add a property which can be used to indicate that there are no devices
on the bus except the ones listed in the devicetree, thus allowing
operating systems to choose to avoid spending time on attempting this
kind of detection of devices which do not exist.

Signed-off-by: Vincent Whitchurch <vincent.whitchurch@axis.com>