schemas: Add an ISA bus schema
ISA bus is still in use in some DTs. Add a basic schema for it so that
'isa' is a documented compatible. There are a few cases of custom ISA
bridges that have a specific binding, so the schema is split between a
bus schema and a generic bridge schema.

Signed-off-by: Rob Herring <robh@kernel.org>