dtschema: Improve error messages for $ref resolution errors
check_schema_refs() may raise an exception or print errors from
process_schema(). In the latter case, the file path is not clear. If the
referring schema had an non-existent file, the referring schema file is
not printed. Fix this by always raising a RefResolutionError exception with
an appropriate message.

Signed-off-by: Rob Herring <robh@kernel.org>