Split DTValidator class into 2 classes
The DTValidator class was originally patterned after jsonschema's
Validator class. That has proven to be problematic as later
deemed non-public APIs were used and then changed. Furthermore as
dtschema has evolved, the checking of schemas and devicetrees have
diverged. Primarily, the DT validation requires preprocessing schemas
and having all the schemas up front to have type information for DTB
decoding.

To better match the uses, split out the schema validation into a
separate DTSchema class and the DTValidator class becomes the class for
validating DTs. As lib.py has grown pretty long, both classes are moved
to their own files.

This serves as prep work to move from jsonschema's soon to be deprecated
RefResolver to referencing library.

It's a big diff, but not easy or really worth trying to do it in small
steps.

Signed-off-by: Rob Herring <robh@kernel.org>