Rework load_schema/process_schemas/set_schema to single 'set_schemas' API
Currently, a different API is exposed depending on whether a pre-processed
schema file or unprocessed schemas files are used. Rework the interface
for loading schemas to be unified whether passing a single pre-processed
schema file or a list of files/directories of schemas.

This ensures the schemas are passed to dtschema module for subsequent calls
to get type or other information from the schemas.

Signed-off-by: Rob Herring <robh@kernel.org>