dtschema: Fix populating pinctrl properties if only 'pinctrl-names' is present
Commit bf775bffa2c1 ("dtschema: Move dropping 'additionalProperties:
true' to after fixup_node_props()") exposed an issue with a schema that
defined 'pinctrl-names' but not 'pinctrl-[0-9]' properties. This was
masked because the automatic properties were being added to a common
schema which was referenced. Prior to that, it was masked by
unevaluatedProperties not being fully working.

Rework the pinctrl check to only look for 'pinctrl-[0-9]' properties
when deciding whether to add pinctrl related properties.

Signed-off-by: Rob Herring <robh@kernel.org>