dtschema: Relax validation of schemas when processed
New meta-schema checks cause existing schemas to have warnings which causes
the entire schema to be omitted when it was previously included. This is
not great because all the new warnings have to be fixed before a new
version of dtschema can be used.

Let's relax the validation to just check against Draft7 schema instead of
the full DT schema. That should be good enough to exclude things which
aren't valid at all.

Signed-off-by: Rob Herring <robh@kernel.org>