dt-validate: Also ignore 'unevaluatedProperties' errors for disabled nodes
In some cases, disabled nodes have misleading 'unevaluatedProperties'
errors. This is caused from 'required' failing in some cases. A simplified
example is a schema like this:

{
    "unevaluatedProperties": false,
    "allOf":[
        {
            "properties": {
                "foo": true,
                "bar": true
            },
            "required": [ "foo" ]
        }
    ]
}

An instance '{ "bar": 1 }' fails with "bar" is unevaluated. This is expected
json-schema behavior.

So we need to ignore 'unevaluatedProperties' as well on disabled nodes. In
the process, let's use the error schema_path rather than parsing the error
message text.

Signed-off-by: Rob Herring <robh@kernel.org>