dtb: Assume empty strings means the type is not string
When there are multiple types including a string type, we test if the
data can be decoded as strings. While empty strings are valid in
general, we only call bytes_to_string() when a property name has
multiple types. In that case, we need to assume empty strings means it
is not a string. Otherwise, values such as a 32-bit 0 looks like 4 empty
strings instead of an integer type.

IIRC, we only needed to allow empty strings when bytes_to_string() was
used in all cases for strings.

Signed-off-by: Rob Herring <robh@kernel.org>