schemas: bootph: Add check for bootph tags in parent nodes
When a node has boot phase tags, the parent node is implicitly included
and a bootph tag is not needed. Add a schema to check for and disallow
bootph tags in both parent and child nodes. It's not the greatest logic
and resulting error message, but that's what works for json-schema.

Signed-off-by: Rob Herring <robh@kernel.org>