meta-schemas: Allow 'not' keyword
Root- and sub-node schemas are supposed to be allowed to use the 'not'
keyword directly to express a schema against which the instance isn't
supposed to be successfully validated. It shall work in the same way as
the allOf, anyOf, oneOf-based validation except the keyword implies a
single sub-schema against which an instance will be anti-evaluated.

Signed-off-by: Serge Semin <Sergey.Semin@baikalelectronics.ru>
Signed-off-by: Rob Herring <robh@kernel.org>