Limit jsonschema version to less than 4.18
The current alpha for jsonschema 4.18 breaks our custom reference
resolution. The API is going to be deprecated as well. Limit the version
until we know jsonschema is fixed or dtschema is updated.

From now on, pin the max version because this wouldn't be the first time
a minor version broke us. :(

Signed-off-by: Rob Herring <robh@kernel.org>