Convert setup.py to pyproject.toml
Use of setup.py has been superseeded by pyproject.toml (and setup.cfg
before that). Now with the latest versions of pip/setuptools in Debian,
there's a warning when running editable install:

DeprecationWarning: pkg_resources is deprecated as an API.

When last looked at pyproject.toml, editable installs weren't supported,
but they are now. So let's move over to the latest and greatest to avoid
this warning.

Signed-off-by: Rob Herring <robh@kernel.org>