validator: Rework property type extraction to save properties without a type
The current checking for properties having a type only checks vendor
specific properties. This mostly works, but misses any property without a
vendor prefix. It also doesn't allow vendor properties defining a type in
one schema and then adding constraints in another. Instead, let's keep
properties which don't have a type. First, we need to track any properties
in the form of 'propname: true' and present a consistent result for no
type. With this, we need to rework our processing of properties and
splitting out of pattern properties.

Signed-off-by: Rob Herring <robh@kernel.org>