meta-schemas: Relax vendor-props meta-schema type checks
Now that we check globally for properties with missing type, there's not
a need to enforce that in the meta-schema. Doing it in the meta-schema
is not very flexible as it doesn't allow defining a property in one
place and adding just constraints in another place except for places we
allow it (if/then/else schemas) or fail to walk (additionalProperties).

Now an error from the global check will cause a schema author to add
'type' or a $ref. Then when one of those are present, the meta-schema
will require 'description'.

Signed-off-by: Rob Herring <robh@kernel.org>