fixups: Skip defaulting minItems/maxItems to the same size on if/then/else schemas
The intent was that if/then/else schemas are not complete, but add to
the top-level schema. A common pattern is increasing the minimum or
decreasing the maximum number of entries for a property. Adding an
implicit minItems/maxItems defeats this. Now that we have the full
schema path available, skip doing this for if/then/else schemas.

Signed-off-by: Rob Herring <robh@kernel.org>