Rework CLI tools to use modern install method
Modern python installs (i.e. using pyproject.toml) want to install CLI
tools by just pointing to a module method and generating the executables
on install. In order to support this, move the installed tools to the
dtschema module directory renaming them to be modules.

This fixes allowing modifications to the tools on editable installs. The
conversion to pyproject.toml inadvertently broke this.

Signed-off-by: Rob Herring <robh@kernel.org>