fixups: Apply int array to matrix fixup even for 1 item
There was a corner case where an array with 1 item was defined:

$ref: /schemas/types.yaml#/definitions/uint32-array
maxItems: 1

While that's not really valid because 1 item is just an uint32, not an
array, the schema didn't work correctly either. With the fixup skipped,
maxItems applied to the outer array (which is always 1 item for arrays)
when the intent was for the inner array.

Since support of 2 forms of array encoding has been dropped, we can also
just drop the maxItems check.

Signed-off-by: Rob Herring <robh@kernel.org>